# Conflicts:
#	gateway/src/main/java/com/patreonshout/rest/WebhookSvc.java

+++ gateway/src/main/java/com/patreonshout/jpa/ListsRepository.java
@@ -2,10 +2,11 @@ package com.patreonshout.jpa;
 
 import com.patreonshout.beans.ListBean;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Spring Data Repository for easy use of CRUD operations on the {@link com.patreonshout.beans.ListBean} object
  */
-public interface ListsRepository extends JpaRepository<ListBean, Long>, ListsRepositoryCustom {
+public interface ListsRepository extends JpaRepository<ListBean, Integer>, ListsRepositoryCustom {
 
 }
